requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
python-binance==1.0.19
pybit==5.6.2
numpy>=1.26.0
pandas>=2.1.0
cryptography>=41.0.0
psycopg2-binary==2.9.9
aiosend>=1.1.0
Flask==3.0.0
anthropic>=0.76.0
openai>=1.12.0
pytest>=7.0.0